move DnsRecords into IntegrationValue
