+++ cont3xt/descriptions.txt
@@ -13,6 +13,7 @@ Card Description - an object of the following
     * otherwise it can be an object with
       * label - the field label to show the user
       * field - the path (it can have dots) to the data for field, if not set the field is the same as name. For table type this will be the path to the array.
+      * path - alternative to field, this is the field path pre-separated into a string array (eg. field: 'foo.bar' is equivalent to path: ['foo', 'bar'])
       * fields - for tables, the list of fields, same format as this
       * type - if not set it is string
         * string - obvious
@@ -24,10 +25,12 @@ Card Description - an object of the following
         * ms - a ms time value
         * seconds - a second time value
         * json - just display raw json, call in JSON.stringify(blah, false, 2)
+        * dnsRecords - display of non-A/AAAA dns records (for use with DNS integration and `path: []`)
       * defang - when true defang the string, change http to hXXp and change . to [.]
       * pivot - when set this field should be added to action menu for table entry that you can replace query with
       * join - with array, display with value as the separator on one line (example single: ', ')
       * fieldRoot - for arrays/tables, the path (it can have dots) from each object in the array to its desired field. Effectively maps the root array of objects to an array of values/sub-objects.
+      * fieldRootPath - alternative to fieldRoot, this is the fieldRoot path pre-separated into a string array (eg. fieldRoot: 'foo.bar' is equivalent to fieldRootPath: ['foo', 'bar'])
       * filterEmpty - for arrays/tables, removes empty (nullish & empty string/array) rows/elements when true (default is true)
       * defaultSortField - with array, sorts the table by this field
       * defaultSortDirection - with array and defaultSortField, sorts the table in this direction ('asc' or 'desc')
